Refactor: 모듈 변경에 따른 의존성 삭제, 컨터버 로직 추가
